pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
aiohttp = "3.9.5"
python = "^3.10"
lavalink = "*"
pillow = "^10.4.0"
numpy = "^2.0.0"
opencv-python = "^4.6.0.66"
pilmoji = { git = "https://github.com/clearlakes/pilmoji" }
"discord.py" = { extras = ["speed"], version = "^2.0.0" }
discord-ext-menus = { git = "https://github.com/Rapptz/discord-ext-menus" }
yt-dlp = "*"
motor = "^3.5.1"
dnspython = "^2.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
